refactor(agents-mobile): remove legacy WebView embed
Use the Expo DOM session surface as the only mobile embed path and delete the old static WebView bundle pipeline.
