+++ packages/agents-mobile/src/lib/embedView.ts
@@ -1,0 +1,1 @@
+export type EmbedViewId = `chat` | `state-explorer`
