--- packages/agents-server-ui/embed.html
@@ -1,16 +1,0 @@
-<!doctype html>
-<html lang="en" data-electric-mobile-embed="true">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no, viewport-fit=cover"
-    />
-    <title>Electric Agents · mobile embed</title>
-    <meta name="color-scheme" content="light dark" />
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/embed/main.tsx"></script>
-  </body>
-</html>
