+++ packages/agents-server-ui/vite.config.ts
@@ -3,7 +3,6 @@ import { defineConfig, type Plugin } from 'vite'
 import react from '@vitejs/plugin-react'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
-import { viteSingleFile } from 'vite-plugin-singlefile'
 
 const PACKAGE_DIR = path.dirname(fileURLToPath(import.meta.url))
 const localNodeModules = (...segments: Array<string>): string =>
@@ -34,7 +33,6 @@ function desktopHtmlMarker(platform: `darwin` | `win32` = `darwin`): Plugin {
 
 export default defineConfig(({ command, mode }) => {
   const desktop = mode === `desktop`
-  const mobileEmbed = mode === `mobile-embed`
   // Desktop *build* serves the bundle via file:// from the Electron
   // app, so assets must be referenced with relative URLs (`./`). The
   // dev server, on the other hand, serves over http and needs an
@@ -43,79 +41,26 @@ export default defineConfig(({ command, mode }) => {
 
   return {
     base: desktop ? (desktopServe ? `/` : `./`) : `/__agent_ui/`,
-    // Mobile embed needs the file-name prefix preserved on CSS module
-    // class names so the embed's mobile-only override sheet (which
-    // targets selectors like `[class*='EntityTimeline_content']`) can
-    // pin point individual rules from the desktop component sheets.
-    // Vite's default production hash (`_[local]_[hash]_[counter]`) drops
-    // the file name; restore it so the embed overrides keep matching
-    // across rebuilds.
-    ...(mobileEmbed
-      ? {
-          css: {
-            modules: {
-              generateScopedName: `[name]_[local]_[hash:base64:5]`,
-            },
-          },
-          // Heavy markdown deps (`mermaid` ~5 MB, `shiki` ~3.8 MB,
-          // `katex` ~600 KB, `@streamdown/math` pulls katex) get
-          // swapped for tiny stubs in `src/embed/stubs/`. Mobile chat
-          // shows code as plain monospace and skips math/diagrams,
-          // which trims the embed bundle from ~13 MB to ~3 MB.
-        }
-      : {}),
     resolve: {
       alias: [
         {
           find: /^@electric-ax\/agents-runtime\/client$/,
           replacement: resolve(__dirname, `../agents-runtime/src/client.ts`),
         },
-        ...(mobileEmbed
-          ? [
-              {
-                find: `mermaid`,
-                replacement: resolve(__dirname, `src/embed/stubs/mermaid.ts`),
-              },
-              {
-                find: /^shiki\/bundle\/web$/,
-                replacement: resolve(__dirname, `src/embed/stubs/shiki.ts`),
-              },
-              {
-                find: /^katex$/,
-                replacement: resolve(__dirname, `src/embed/stubs/katex.ts`),
-              },
-              {
-                find: /^@streamdown\/math$/,
-                replacement: resolve(
-                  __dirname,
-                  `src/embed/stubs/streamdown-math.ts`
-                ),
-              },
-            ]
-          : []),
+        { find: `react`, replacement: localNodeModules(`react`) },
+        { find: `react-dom`, replacement: localNodeModules(`react-dom`) },
+        {
+          find: `react/jsx-runtime`,
+          replacement: localNodeModules(`react`, `jsx-runtime.js`),
+        },
+        {
+          find: `react/jsx-dev-runtime`,
+          replacement: localNodeModules(`react`, `jsx-dev-runtime.js`),
+        },
       ],
-    },
-    plugins: [
-      react(),
-      ...(desktop ? [desktopHtmlMarker()] : []),
-      // Mobile embed inlines all JS + CSS into a single HTML so the
-      // React Native side can ship one self-contained string.
-      ...(mobileEmbed
-        ? [viteSingleFile({ removeViteModuleLoader: true })]
-        : []),
-    ],
-    resolve: {
-      alias: {
-        react: localNodeModules(`react`),
-        'react-dom': localNodeModules(`react-dom`),
-        'react/jsx-runtime': localNodeModules(`react`, `jsx-runtime.js`),
-        'react/jsx-dev-runtime': localNodeModules(
-          `react`,
-          `jsx-dev-runtime.js`
-        ),
-      },
       dedupe: [`react`, `react-dom`],
     },
+    plugins: [react(), ...(desktop ? [desktopHtmlMarker()] : [])],
     optimizeDeps: {
       include: [
         `react`,
@@ -125,25 +70,8 @@ export default defineConfig(({ command, mode }) => {
       ],
     },
     build: {
-      outDir: mobileEmbed
-        ? `dist-mobile-embed`
-        : desktop
-          ? `dist-desktop`
-          : `dist`,
+      outDir: desktop ? `dist-desktop` : `dist`,
       emptyOutDir: true,
-      // Use `embed.html` as the lone entry for the mobile-embed bundle
-      // so the regular `index.html` (workspace shell) is excluded.
-      ...(mobileEmbed
-        ? {
-            rollupOptions: {
-              input: { index: resolve(__dirname, `embed.html`) },
-            },
-            // Keep things in a single chunk for the singlefile inliner.
-            assetsInlineLimit: 100 * 1024 * 1024,
-            chunkSizeWarningLimit: 100 * 1024,
-            cssCodeSplit: false,
-          }
-        : {}),
     },
   }
 })
